cli-agent/requirements.txt
azure-identity>=1.12.0
azure-mgmt-containerservice>=20.0.0
azure-mgmt-resource>=21.1.0
click>=8.0.0
kubernetes>=24.2.0
pyyaml>=6.0
requests>=2.28.0
rich>=12.0.0
openai>=1.0.0
anthropic>=0.3.0